* Fixed inability to use gradients properly. * Re-added ability to toggle the use of & symbols in clan names for manual coloring. * Optimized and switched out deprecated task scheduling usages. * Added usage of "/clan color reset" will remove any coloring or gradient or clan might have * Added new claim charging configuration enchantments credit to grzybeek * Fixed alias conversions for /clan help command list. * Fixed /c map not using the right chunk claiming methods. * Fixed /c bio with no arguments giving internal error. * Added new prevention for invalid claims on caching. * Finalized Roundtable (Now utilizes all configuration) * Optimized placeholder translations. * Fixed /c territory chunk visualizer task * Fixed improper constructor for Teleport.Impl.java * Optimized start procedure * Fixed password change prompt and added new relative configurable message response. * Added new permission handler allowing for use of wildcards (examples in Messages.yml) * Replaced old ClanRosterElement.java with up front collection. * Say hello to the complete new gui configuration! We've had a few complaints about gui customization, so we're happy to announce it's all 100% customizable now! With hundreds of gui configuration combinations go wild and make it look however you want :) * Added the ability to customize the /c invite command into depth * Added /c block message response. * Added ability to disable /c bank command (new config option) * There was a large request to add configuration for the placeholder that returns "N/A" when empty so this has also been added. * Added new addon "Traits" to be nostalgic of Clans [Classes] from early clans free, adds (for now) 3 classes (Assassin, Brute & Tamer) and 4 abilities (Selective Kill, Full Strength, Pick Pocket, Summon Wolves) you can level each ability to make the reward greater. * Fixed possible persistence problems with claim loading due to invisible mapping (A bug i dealt with while debugging not sure if it was just my side but its resolved, i got no complaints about it)

+++ ClansPro/src/main/java/com/github/sanctum/clans/bridge/internal/vaults/VaultMenu.java
@@ -29,52 +29,53 @@ public class VaultMenu extends Menu {
 	public VaultMenu(String clanName) {
 		super(ClansAPI.getInstance().getPlugin(), clanName, Rows.SIX, Type.SINGULAR, Property.SHAREABLE, Property.CACHEABLE);
 		this.c = ClansAPI.getInstance().getClanManager().getClan(ClansAPI.getInstance().getClanManager().getClanID(clanName));
-		addElement(new InventoryElement.Shared(clanName, this));
-		this.key = clanName;
-		this.close = close -> {
-			FileManager manager = ClanAddon.getAddon(VaultsAddon.class).getFile(FileType.JSON, "data");
-			Node keys = manager.getRoot().getNode(c.getId().toString());
-			keys.delete();
-			int key = 0;
-			DataTable table = DataTable.newTable();
-			for (ItemStack item : close.getMain().getContents()) {
-				if (item != null && item.getType() != Material.AIR) {
-					table.set(c.getId() + "." + key, item);
-					key++;
+		if (this.c != null) {
+			addElement(new InventoryElement.Shared(clanName, this));
+			this.key = clanName;
+			this.close = close -> {
+				FileManager manager = ClanAddon.getAddon(VaultsAddon.class).getFile(FileType.JSON, "data");
+				Node keys = manager.getRoot().getNode(c.getId().toString());
+				keys.delete();
+				int key = 0;
+				DataTable table = DataTable.newTable();
+				for (ItemStack item : close.getMain().getContents()) {
+					if (item != null && item.getType() != Material.AIR) {
+						table.set(c.getId() + "." + key, item);
+						key++;
+					}
 				}
-			}
-			manager.write(table, true);
-		};
+				manager.write(table, true);
+			};
 
-		this.click = click -> {
-			VaultInteractEvent event = ClanVentBus.call(new VaultInteractEvent(c, click.getElement(), click.getParent().getElement(), click.getAttachment(), click.getAction(), click.getClickType(), click.getParent().getParent().getElement().getViewers()));
-			if (event.isCancelled()) {
-				click.setCancelled(true);
-			}
-		};
-		if (oldExists(c.getId())) {
-			for (ItemStack it : getInventoryContentsOld(Check.forNull(c.getId(), "Clan '" + clanName + "' not found!"))) {
-				if (Objects.equals(it, null)) {
-					getInventory().getElement().addItem(new ItemStack(Material.AIR));
-				} else {
-					getInventory().getElement().addItem(it);
+			this.click = click -> {
+				VaultInteractEvent event = ClanVentBus.call(new VaultInteractEvent(c, click.getElement(), click.getParent().getElement(), click.getAttachment(), click.getAction(), click.getClickType(), click.getParent().getParent().getElement().getViewers()));
+				if (event.isCancelled()) {
+					click.setCancelled(true);
+				}
+			};
+			if (oldExists(c.getId())) {
+				for (ItemStack it : getInventoryContentsOld(Check.forNull(c.getId(), "Clan '" + clanName + "' not found!"))) {
+					if (Objects.equals(it, null)) {
+						getInventory().getElement().addItem(new ItemStack(Material.AIR));
+					} else {
+						getInventory().getElement().addItem(it);
+					}
+				}
+				c.removeValue("vault");
+			} else {
+				for (ItemStack it : getInventoryContentNew(Check.forNull(c.getId(), "Clan '" + clanName + "' not found!"))) {
+					if (Objects.equals(it, null)) {
+						getInventory().getElement().addItem(new ItemStack(Material.AIR));
+					} else {
+						getInventory().getElement().addItem(it);
+					}
 				}
 			}
-			c.removeValue("vault");
-		} else {
-			for (ItemStack it : getInventoryContentNew(Check.forNull(c.getId(), "Clan '" + clanName + "' not found!"))) {
-				if (Objects.equals(it, null)) {
-					getInventory().getElement().addItem(new ItemStack(Material.AIR));
-				} else {
-					getInventory().getElement().addItem(it);
-				}
+			try {
+				registerController();
+			} catch (Exception ignored) {
 			}
 		}
-		try {
-			registerController();
-		} catch (Exception ignored) {
-		}
-
 	}
 
 	boolean oldExists(HUID id) {
